# docs/swagger.yaml
openapi: 3.0.3
info:
  title: 'Personal Expense Tracker API with Chatbot Analyst'
  description: |
    A RESTful API for tracking personal expenses with integrated chatbot analysis.

    ## Authentication
    The API uses session-based authentication with HTTP-only cookies.
    - Register or login to receive a `session_token` cookie (7-day expiry)
    - Session tokens are stored server-side as SHA-256 hashes and are revocable
    - Include the cookie in subsequent requests
    - Use the `/api/me` endpoint to verify authentication status

    ## Rate limiting
    - `POST /api/auth/login`: 10 attempts / 15 min per email AND 30 / 15 min per IP
      (successful login resets the email bucket)
    - `POST /api/auth/register`: 10 / 15 min per IP
    - `POST /api/chatbot`: 50 messages / user / UTC day
    - Limited endpoints return 429; auth endpoints include a `Retry-After` header

    ## Conventions
    - **Money:** all amounts are integers in **cents** (e.g. `4999` = $49.99)
    - **IDs:** MongoDB ObjectIds — 24-character hexadecimal strings
    - **Ownership:** on single-resource endpoints, `404` means "not found *or*
      not owned by the requesting user" — the API never reveals whether a
      resource exists for someone else
    - **Dates:** accepted as `YYYY-MM-DD` or full ISO 8601 datetime strings;
      always returned as ISO 8601 datetimes
    - **Errors:** every error response uses the same envelope:
      `{ success: false, error, details? }`
...
      this API

...
servers:
...
paths:
...
  /expenses:
    get:
      tags: [Client]
      summary: Get expenses dashboard
      description: Returns the expenses dashboard for authenticated users.
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Expenses dashboard rendered successfully
          content:
            text/html:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/expenses:
    get:
      tags: [Expenses]
      summary: List expenses
      description: |
        Returns a paginated list of the authenticated user's expenses, ordered
        by date (descending). Optionally filter by month and category.
      security:
        - sessionAuth: []
      parameters:
        - name: month
          in: query
          required: false
          description: Filter to a single month (YYYY-MM)
          schema:
            type: string
            pattern: '^\d{4}-(0[1-9]|1[0-2])$'
            example: '2025-06'
        - name: categoryId
          in: query
          required: false
          description: Filter to one of the user's categories
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Expenses retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Expenses]
      summary: Create an expense
      description: |
        Creates an expense record for the authenticated user.

        If `categoryId` is provided it must reference one of *this user's*
        categories — otherwise the request is rejected with 400 (this
        prevents linking another user's category).

        Creating an expense also materializes any due recurring income
        entries (see Conventions).
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseCreateRequest'
            example:
              name: Weekly groceries
              amount: 4999
              date: '2025-06-10'
              categoryId: 507f1f77bcf86cd799439013
      responses:
        '201':
          description: Expense created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: HTTP-only session cookie obtained from login or registration
  schemas:
...
    ErrorResponse:
      type: object
      description: |
        Canonical error envelope — every error response in the API uses this
        shape. `details` (Zod issue objects) is present only on validation
        failures.
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Internal Server Error
        details:
          type: array
          description: Present on validation failures (Zod issue objects)
          items:
            $ref: '#/components/schemas/ZodIssue'
        resetAt:
          type: string
          format: date-time
          description: Only on chatbot 429s — when the daily limit resets
      required:
        - success
        - error
    ZodIssue:
      type: object
      description: Validation issue produced by Zod
      properties:
        code:
          type: string
          example: invalid_format
        message:
          type: string
          example: Invalid email address
        path:
          type: array
          items:
            type: string
          example:
            - email
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        total:
          type: integer
          description: Total number of matching records
          example: 42
        totalPages:
          type: integer
          example: 3
...
    Expense:
      type: object
      properties:
        id:
          type: string
          description: MongoDB ObjectId
          example: 507f1f77bcf86cd799439021
        date:
          type: string
          format: date-time
          description: When the expense occurred
          example: '2025-06-10T00:00:00.000Z'
        name:
          type: string
          example: Weekly groceries
        amount:
          type: integer
          description: Amount in cents
          example: 4999
        categoryId:
          type: string
          nullable: true
          description: Category ObjectId, or null if uncategorized
          example: 507f1f77bcf86cd799439013
        category:
          type: object
          nullable: true
          description: The expense's category, or null if uncategorized
          properties:
            id:
              type: string
              example: 507f1f77bcf86cd799439013
            name:
              type: string
              example: groceries
        userId:
          type: string
          description: Owning user's ObjectId
          example: 507f1f77bcf86cd799439011
      required:
        - id
        - date
        - name
        - amount
        - userId
    ExpenseCreateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Weekly groceries
        amount:
          type: integer
          minimum: 1
          description: Amount in cents
          example: 4999
        date:
          type: string
          description: 'YYYY-MM-DD or full ISO 8601 datetime'
          example: '2025-06-10'
        categoryId:
          type: string
          description: Optional. Must reference a category owned by the requesting user.
          example: 507f1f77bcf86cd799439013
      required:
        - name
        - amount
        - date
...
    ExpenseListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Expense'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
  responses:
    Unauthorized:
      description: Authentication required or invalid session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Unauthorized
    BadRequest:
      description: |
        Validation failed (with `details`), a referenced resource is invalid
        (e.g. a categoryId owned by another user), or a query/path parameter
        is malformed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              summary: Validation failure with details
              value:
                success: false
                error: Validation failed
                details:
                  - code: invalid_format
                    message: Invalid email address
                    path:
                      - email
            invalidReference:
              summary: Referenced resource not owned by the user
              value:
                success: false
                error: Category not found
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Internal Server Error